+++ libs/database/migrations/0028_far_chronomancer.sql
@@ -1,0 +1,6 @@
+ALTER TABLE "subscriptions" DROP CONSTRAINT "subscriptions_user_id_users_id_fk";
+--> statement-breakpoint
+ALTER TABLE "sent_notifications" DROP CONSTRAINT "sent_notifications_subscription_id_subscriptions_id_fk";
+--> statement-breakpoint
+ALTER TABLE "subscriptions" ADD CONSTRAINT "subscriptions_user_id_users_id_fk" FOREIGN KEY ("user_id") REFERENCES "public"."users"("id") ON DELETE cascade ON UPDATE no action;--> statement-breakpoint
+ALTER TABLE "sent_notifications" ADD CONSTRAINT "sent_notifications_subscription_id_subscriptions_id_fk" FOREIGN KEY ("subscription_id") REFERENCES "public"."subscriptions"("id") ON DELETE cascade ON UPDATE no action;